Rename package engines->engine to match internal structure
