+++ examples/plugins/test-plugin/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	pluginapi "github.com/sylabs/singularity/pkg/plugin"
-	singularity "github.com/sylabs/singularity/pkg/runtime/engines/singularity/config"
+	singularity "github.com/sylabs/singularity/pkg/runtime/engine/singularity/config"
 )
 
 // Plugin is the only variable which a plugin MUST export. This symbol is accessed
