+++ internal/pkg/runtime/engine/singularity/engine_linux.go
@@ -9,7 +9,7 @@ import (
 	"github.com/sylabs/singularity/internal/pkg/runtime/engine"
 	"github.com/sylabs/singularity/internal/pkg/runtime/engine/config"
 	"github.com/sylabs/singularity/internal/pkg/runtime/engine/singularity/rpc/server"
-	singularityConfig "github.com/sylabs/singularity/pkg/runtime/engines/singularity/config"
+	singularityConfig "github.com/sylabs/singularity/pkg/runtime/engine/singularity/config"
 )
 
 // EngineOperations is a Singularity runtime engine that implements engine.Operations.
