+++ mlocal/frags/build_cli.mk
@@ -54,7 +54,7 @@ config_INSTALL := $(DESTDIR)$(SYSCONFDIR)/singularity/singularity.conf
 # override this to empty to avoid merging old configuration settings
 old_config := $(config_INSTALL)
 
-$(config): $(singularity_build_config) $(SOURCEDIR)/etc/conf/gen.go $(SOURCEDIR)/etc/conf/singularity.conf $(SOURCEDIR)/pkg/runtime/engines/singularity/config/config.go
+$(config): $(singularity_build_config) $(SOURCEDIR)/etc/conf/gen.go $(SOURCEDIR)/etc/conf/singularity.conf $(SOURCEDIR)/pkg/runtime/engine/singularity/config/config.go
 	@echo " GEN $@`if [ -n "$(old_config)" ]; then echo " from $(old_config)"; fi`"
 	$(V)$(GO) run $(GO_MODFLAGS) $(GO_GCFLAGS) $(GO_ASMFLAGS) $(SOURCEDIR)/etc/conf/gen.go \
 		$(SOURCEDIR)/etc/conf/singularity.conf $(old_config) $(config)
